use creation and update timestamps, allow migration to be handled by spring boot and confirm question exists before attempting delete

+++ src/main/java/pro/arcodeh/msgoba_2002_server/question/QuestionService.java
@@ -56,6 +56,12 @@ public class QuestionService {
     }
 
     public void deleteQuestion(UUID questionId) {
+        boolean questionExists = this.questionRepository.existsById(questionId);
+
+        if(questionExists) {
+            throw new NotFoundException("Question not found. Deletion failed.");
+        }
+
         this.questionRepository.deleteById(questionId);
     }
 }
